fix(Remove Detailed posts): Change the settings label from "detailed" to "related"
"Tweet Detail Related Tweets" was mistakenly abbreviated to "Detailed".
"Related" is more accurate word to describe "Discover more" section.

However, the name of the patch is not changed in this commit.
Because renaming the patch name might affect in manager/builder that's use this patch name.
Variable names and string keys are also not changed as per the patch name.