Remove unused patches from ReVanced
My understanding is that piko doesn't use these patches.
If you copied this for future use, please let me know so I can revert it.

However, in my opinion, it's better to copy these when they are actually needed for piko.
Copying them now will make it burdensome to reflect changes in the official ReVanced.